+++ assets/javascripts/kitten/index.js
@@ -9,6 +9,8 @@ export { VisuallyHidden } from './components/accessibility/visually-hidden'
 export { Accordeon } from './components/accordeon'
 
 // Atoms
+export { Loader } from './components/atoms/loader'
+export { LoaderWithParagraph } from './components/atoms/loader-with-paragraph'
 export { Tag } from './components/atoms/tag'
 
 // Avatar
@@ -168,10 +170,6 @@ export { CheckList } from './components/lists/check-list'
 export { NavList } from './components/lists/nav-list'
 export { TagList } from './components/lists/tag-list'
 export { Timeline } from './components/lists/timeline'
-
-// Loaders
-export { LoaderWithParagraph } from './components/loaders/loader-with-paragraph'
-export { Loader } from './components/loaders/loader'
 
 // Menus
 export { HeaderMenu } from './components/menus/header-menu'